requirements.txt
Flask==1.1.1
CTRegisterMicroserviceFlask==0.3.0
requests==2.23.0
httmock==1.2.6
earthengine-api==0.1.236
redis==2.10.6
google-cloud-storage==1.18.0
urllib3==1.25.8
oauth2client==3.0.0
python-dotenv==0.14.0
redis==2.10.6
gevent==20.9.0